requirements.txt
python-jenkins==0.4.16
